Refactor simple-standard-scenarios tests to use cached data
